Move common components together
