+++ static/app/views/performance/mobile/common/components/deviceClassSelector.tsx
