+++ static/app/views/performance/mobile/common/components/screensTemplate.tsx
