+++ static/app/views/performance/mobile/common/components/spanSamplesPanel.tsx
@@ -16,8 +16,8 @@ import {normalizeUrl} from 'sentry/utils/withDomainRequired';
 import DetailPanel from 'sentry/views/insights/common/components/detailPanel';
 import {useReleaseSelection} from 'sentry/views/insights/common/queries/useReleases';
 import type {ModuleName} from 'sentry/views/insights/types';
+import {SpanSamplesContainer} from 'sentry/views/performance/mobile/common/components/spanSamplesPanelContainer';
 import useCrossPlatformProject from 'sentry/views/performance/mobile/common/queries/useCrossPlatformProject';
-import {SpanSamplesContainer} from 'sentry/views/performance/mobile/components/spanSamplesPanelContainer';
 
 type Props = {
   groupId: string;
