+++ static/app/views/performance/mobile/ui/screenSummary/index.tsx
@@ -13,8 +13,8 @@ import {useLocation} from 'sentry/utils/useLocation';
 import useRouter from 'sentry/utils/useRouter';
 import {ReleaseComparisonSelector} from 'sentry/views/insights/common/components/releaseSelector';
 import {ModuleName, SpanMetricsField} from 'sentry/views/insights/types';
-import {SamplesTables} from 'sentry/views/performance/mobile/components/samplesTables';
-import {SpanSamplesPanel} from 'sentry/views/performance/mobile/components/spanSamplesPanel';
+import {SpanSamplesPanel} from 'sentry/views/performance/mobile/common/components/spanSamplesPanel';
+import {SamplesTables} from 'sentry/views/performance/mobile/common/components/tables/samplesTables';
 import {SpanOperationTable} from 'sentry/views/performance/mobile/ui/screenSummary/spanOperationTable';
 import {ModulePageProviders} from 'sentry/views/performance/modulePageProviders';
 import {useModuleBreadcrumbs} from 'sentry/views/performance/utils/useModuleBreadcrumbs';
